pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "haystack_threshold_node"
version = "0.0.1"
authors = [
  { name="recrudesce", email="recrudesce@gmail.com" },
]
description = "This component filters documents based on a threshold percentage, ensuring only the documents above the threshold get passed down the pipeline"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
  "farm-haystack",
]

[project.urls]
"Homepage" = "https://github.com/recrudesce/haystack_threshold_node"
"Bug Tracker" = "https://github.com/recrudesce/haystack_threshold_node/issues"

